chore: add register-adoption-form hooks folder
