+++ apps/mobile/src/components/register-adoption-form/hooks/use-form-validation.hook.test.ts
@@ -2,10 +2,10 @@ import { act, renderHook } from '@testing-library/react-hooks';
 import { useToast } from 'native-base';
 import { useFormContext } from 'react-hook-form';
 
-import { mountErrorMessage, useFormValidation } from './adoption-form.hooks';
-import useLocale from '../../shared/hooks/use-locale';
+import { mountErrorMessage, useFormValidation } from './use-form-validation.hook';
+import useLocale from '../../../shared/hooks/use-locale';
 
-jest.mock('../../shared/hooks/use-locale');
+jest.mock('../../../shared/hooks/use-locale');
 jest.mock('native-base');
 jest.mock('react-hook-form');
