+++ apps/mobile/src/components/register-adoption-form/hooks/use-multi-step-navigation.hook.test.ts
@@ -1,13 +1,27 @@
 import { act, renderHook } from '@testing-library/react-hooks';
 
-import { stepsLibrary } from './adoption-form.constants';
-import { getStepsByOrder, useMultiStepNavigation } from './adoption-form.hooks';
-import { AdoptionSteps } from './adoption-form.types';
+import { getStepsByOrder, useMultiStepNavigation } from './use-multi-step-navigation.hook';
+import { stepsLibrary } from '../adoption-form.constants';
+import { AdoptionSteps } from '../adoption-form.types';
 
 jest.mock('native-base', () => ({
   ...jest.requireActual('native-base'),
   useToast: jest.fn(),
 }));
+
+describe('getStepsByOrder', () => {
+  it('returns step ids ordered', () => {
+    expect(getStepsByOrder(stepsLibrary)).toStrictEqual({
+      0: AdoptionSteps.PetName,
+      1: AdoptionSteps.PetBreed,
+      2: AdoptionSteps.PetType,
+      3: AdoptionSteps.PetAge,
+      4: AdoptionSteps.PetGender,
+      5: AdoptionSteps.PetSize,
+      6: AdoptionSteps.PetObservations,
+    });
+  });
+});
 
 describe('useMultiStepNavigation', () => {
   describe('when the current step is PetObservations', () => {
@@ -49,20 +63,6 @@ describe('useMultiStepNavigation', () => {
         expect(result.current.isFirstStep).toBeFalsy();
         expect(result.current.activeStep).toBe(AdoptionSteps.PetSize);
       });
-    });
-  });
-});
-
-describe('getStepsByOrder', () => {
-  it('returns step ids ordered', () => {
-    expect(getStepsByOrder(stepsLibrary)).toStrictEqual({
-      0: AdoptionSteps.PetName,
-      1: AdoptionSteps.PetBreed,
-      2: AdoptionSteps.PetType,
-      3: AdoptionSteps.PetAge,
-      4: AdoptionSteps.PetGender,
-      5: AdoptionSteps.PetSize,
-      6: AdoptionSteps.PetObservations,
     });
   });
 });
